Update docs split tests dnsmessage add more tests
